+++ spending-ui/src/components/batches/batches.component.jsx
@@ -1,36 +1,13 @@
-/* eslint max-len: 0 */
-/* eslint no-unused-vars: 0 */
-
+import React from "react";
 import {useContext, useEffect, useState} from "react";
 
-import Collapsible from 'react-collapsible';
-import BootstrapTable from 'react-bootstrap-table-next';
-
-import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css';
-import {nanoid} from 'nanoid';
-
-import {StaticDataContext} from "../../contexts/static_data.context";
-
-import paginationFactory from "react-bootstrap-table2-paginator";
-import React from "react";
-import ReactDOM from "react-dom";
-import Select from "react-select";
-import { Row } from "react-bootstrap";
-import filterFactory, {
-    Comparator,
-    customFilter,
-    FILTER_TYPES
-} from "react-bootstrap-table2-filter";
-import { contextMenu, Item, Menu, Separator, Submenu } from "react-contexify";
 import "react-contexify/dist/ReactContexify.css";
-import cellEditFactory from 'react-bootstrap-table2-editor';
-
-import FormInput from "../form-input/form-input.component";
-import Button from "../button/button-component";
+import { Row } from "react-bootstrap";
+import TableBaseComponent from '../table-base/table-base.component.jsx';
+import {StaticDataContext} from "../../contexts/static_data.context";
 
 const BatchesComponent = () => {
     const {batches} = useContext(StaticDataContext);
-    const [activeRow, setActiveRow] = useState(0);
     const [isLoaded, setIsLoaded] = useState(false);
 
     const columns = [];
@@ -47,35 +24,6 @@ const BatchesComponent = () => {
         }
     }, [batches]);
 
-    const rowStyle = (row) => {
-        if (row === activeRow) {
-            return {
-                backgroundColor: "lightcyan",
-                border: "solid 2px grey",
-                color: "purple"
-            };
-        }
-    };
-
-    const showContext = (event, row) => {
-        console.log("showContext: ", event);
-        setActiveRow(row);
-        event.preventDefault();
-        contextMenu.show({
-            id: "context-menu",
-            event: event
-        });
-    };
-
-    const rowEvents = {
-        onClick: (e, row, index) => {
-            setActiveRow(row);
-        },
-        onContextMenu: (e, row, index) => {
-            showContext(e, row);
-        }
-    };
-
     if(isLoaded) {
         return (
             <div>
@@ -83,28 +31,7 @@ const BatchesComponent = () => {
                     <h1>Batches</h1>
                 </Row>
                 <Row>
-                    <BootstrapTable
-                        keyField='id'
-                        data={batches}
-                        columns={columns}
-                        rowEvents={rowEvents}
-                        rowStyle={rowStyle}
-                    />
-                    <Menu id="context-menu" theme='dark'>
-                        {activeRow && (
-                            <>
-                                <Item className="text-center">Header row {activeRow.id}</Item>
-                                <Separator/>
-                                {["Google", "Apple"].includes("Google") && (
-                                    <Submenu label="Contact" arrow=">">
-                                        <Item>Phone</Item>
-                                        <Item>Email</Item>
-                                    </Submenu>
-                                )}
-                                <Item disabled={true}>Add to Cart</Item>
-                            </>
-                        )}
-                    </Menu>
+                    <TableBaseComponent columns={columns} data={batches} keyField='id'/>
                 </Row>
             </div>
         )
